+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## 1.1.0 - 2016-07-17
+### Added
+- [#7](https://github.com/krux/stubborn/issues/7) Support retry on Either
+- [#9](https://github.com/krux/stubborn/issues/9) Support exponential backoff
+
 ## 1.0.0 - 2016-07-07
 ### Changed
-- First public release with fixed and expoential backoff and jitter retry policy
+- First public release with fixed and exponential backoff and jitter retry policy